refactor: move all client-only pages to pages/ from app/
