--- app/layout.tsx
@@ -1,29 +1,0 @@
-import { AppRouterCacheProvider } from '@mui/material-nextjs/v14-appRouter';
-import type { Metadata } from 'next';
-import { Inter } from 'next/font/google';
-import './globals.css';
-import ScrollToTop from '@/components/ScrollToTop/ScrollToTop';
-
-const inter = Inter({ subsets: ['latin'] });
-
-const metadata: Metadata = {
-	title: 'Majeggstics',
-	description: 'Generated by create next app',
-};
-
-export default function RootLayout({
-	children,
-}: Readonly<{
-	children: React.ReactNode;
-}>) {
-	return (
-		<html lang="en">
-			<AppRouterCacheProvider>
-				<body className={inter.className}>
-					{children}
-					<ScrollToTop />
-				</body>
-			</AppRouterCacheProvider>
-		</html>
-	);
-}
