+++ pages/contract-boost-calculator/constants.ts
