+++ pages/guide/images/farmReadyToReceiveCBs.png
