+++ pages/initiator-tools/notin-message-generator.tsx
@@ -1,7 +1,7 @@
 // Currently doesn't support Wonky's "(from timeslot xx)"
 'use client';
-
 import copy from 'copy-to-clipboard';
+import Head from 'next/head';
 import { Fragment, useState, useCallback, useMemo } from 'react';
 import ToastMessage from '@/components/ToastMessage';
 import { Timeslot, useExtractNotins, type UserSpec } from '@/components/initiator-tools';
@@ -51,42 +51,47 @@ export default function NotInMessageGeneratorPage() {
 	const notins = useExtractNotins(notInMessage);
 
 	return (
-		<div style={{ margin: '2rem 1rem' }}>
-			<ToastMessage />
-			<h1>NotIn Message Generator</h1>
-			<p style={{ display: 'flex', flexDirection: 'column' }}>
-				<label htmlFor="#notInMessage">Not in message from Wonky</label>
-				<textarea
-					id="notInMessage"
-					name="notInMessage"
-					onChange={handleNotInMessageChange}
-					rows={10}
-					style={{ margin: '1rem 0' }}
-					value={notInMessage}
-				/>
-			</p>
+		<>
+			<Head>
+				<title>Notin Message Generator</title>
+			</Head>
+			<div style={{ margin: '2rem 1rem' }}>
+				<ToastMessage />
+				<h1>NotIn Message Generator</h1>
+				<p style={{ display: 'flex', flexDirection: 'column' }}>
+					<label htmlFor="#notInMessage">Not in message from Wonky</label>
+					<textarea
+						id="notInMessage"
+						name="notInMessage"
+						onChange={handleNotInMessageChange}
+						rows={10}
+						style={{ margin: '1rem 0' }}
+						value={notInMessage}
+					/>
+				</p>
 
-			{Object.entries(notins).map(([timeslotEmoji, slotNotins]) => (
-				<Fragment key={`header-${timeslotEmoji}`}>
-					<h4>{Timeslot.fromEmoji(timeslotEmoji)!.format('header')}</h4>
-					<div
-						style={{
-							display: 'grid',
-							gridTemplateColumns: 'max-content repeat(2, min-content) auto',
-							gridColumnGap: '0.5rem',
-							gridRowGap: '1rem',
-						}}
-					>
-						{slotNotins.map(({ users, threadUrl }, userIndex) => (
-							<NotIn
-								key={`${timeslotEmoji}-${userIndex}`}
-								user={users.map((user: UserSpec) => user.combinedIdentifier).join(', ')}
-								{...{ timeslotEmoji, threadUrl }}
-							/>
-						))}
-					</div>
-				</Fragment>
-			))}
-		</div>
+				{Object.entries(notins).map(([timeslotEmoji, slotNotins]) => (
+					<Fragment key={`header-${timeslotEmoji}`}>
+						<h4>{Timeslot.fromEmoji(timeslotEmoji)!.format('header')}</h4>
+						<div
+							style={{
+								display: 'grid',
+								gridTemplateColumns: 'max-content repeat(2, min-content) auto',
+								gridColumnGap: '0.5rem',
+								gridRowGap: '1rem',
+							}}
+						>
+							{slotNotins.map(({ users, threadUrl }, userIndex) => (
+								<NotIn
+									key={`${timeslotEmoji}-${userIndex}`}
+									user={users.map((user: UserSpec) => user.combinedIdentifier).join(', ')}
+									{...{ timeslotEmoji, threadUrl }}
+								/>
+							))}
+						</div>
+					</Fragment>
+				))}
+			</div>
+		</>
 	);
 }
